% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bf11}
\alias{bf11}
\title{3d topological surface measurements for a breech face impression of a cartridge case of the Todd Weller study}
\format{a list}
\usage{
bf11
}
\description{
need more info
}
\keyword{datasets}

