% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p.R
\docType{data}
\name{br411}
\alias{br411}
\title{3d topological surface measurements for one land of a bullet from the Hamby study}
\format{a list}
\usage{
br411
}
\description{
Some more info - not sure at the moment which bullet this is. Describe structure.
}
\keyword{datasets}
