% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{bulletGetMaxCMS}
\alias{bulletGetMaxCMS}
\title{Identify the number of maximum CMS between two bullet lands}
\usage{
bulletGetMaxCMS(lof1, lof2, column = "resid", span = 35)
}
\arguments{
\item{lof1}{dataframe of smoothed first signature}

\item{lof2}{dataframe of smoothed second signature}

\item{column}{The column which to smooth}

\item{span}{positive number  for the smoothfactor to use for assessing peaks.}
}
\value{
list of matching parameters, data set of the identified striae, and the aligned data sets.
}
\description{
Identify the number of maximum CMS between two bullet lands
}
