% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{bulletSmooth}
\alias{bulletSmooth}
\title{Smooth the surface of a bullet}
\usage{
bulletSmooth(data, span = 0.03, limits = c(-5, 5))
}
\arguments{
\item{data}{data frame as returned by the function \code{processBullets}}

\item{span}{width of the smoother, defaults to 0.03}

\item{limits}{vector of the form c(min, max). Results will be limited to be between these values.}
}
\value{
data frame of the same form as the input extended by the vector l30 for the smooth.
}
\description{
Smooth the surface of a bullet
}
