% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley.R
\name{chumbley}
\alias{chumbley}
\title{Chumbley test score}
\usage{
chumbley(b1, b2, window, reps = 3)
}
\arguments{
\item{b1}{dataframe}

\item{b2}{dataframe}

\item{window}{width of the window (in indices) to consider for matching}

\item{reps}{number of replicates to use in the evaluation}
}
\description{
Chumbley test score
}
\examples{
data(br411)
b1 <- get_crosscut(x = 250, bullet=br411)
b2 <- get_crosscut(x = 150, bullet = br411)
b3 <- get_crosscut(x = 10, bullet=br411)
b1.gr <- b1 \%>\% get_grooves(smoothfactor=30)
b2.gr <- b2 \%>\% get_grooves()
b3.gr <- b3 \%>\% get_grooves()
# check that the grooves are actually found:
b1.gr$plot
b2.gr$plot
# get signatures
b1 <- fit_loess(b1, b1.gr)$data
b2 <- fit_loess(b2, b2.gr)$data
b3 <- fit_loess(b3, b3.gr)$data
# matched correlations
get_cor(b1, b2, window = 100, b1.left = 800, lag = match$lag)
get_cor(b1, b2, window = 100, b1.left = 1000, lag = match$lag)
get_cor(b1, b2, window = 100, b1.left = 1200, lag = match$lag)
# random correlations
get_cor(b1, b2, window = 100, b1.left = 800, lag = 100)
get_cor(b1, b2, window = 100, b1.left = 1000, lag = -300)
get_cor(b1, b2, window = 100, b1.left = 1200, lag = -500)

chumbley(b1, b2, window=150, reps=5)

match13 <- get_lag_max_R(b1, b3, window = 100, b1.left = 450)
# matched correlations
get_cor(b1, b3, window = 100, b1.left = 800, lag = match13$lag)
get_cor(b1, b3, window = 100, b1.left = 1000, lag = match13$lag)
get_cor(b1, b3, window = 100, b1.left = 1200, lag = match13$lag)
# random correlations
get_cor(b1, b3, window = 100, b1.left = 800, lag = 100)
get_cor(b1, b3, window = 100, b1.left = 1000, lag = 300)
get_cor(b1, b3, window = 100, b1.left = 1200, lag = 500)

chumbley(b1, b3, window=100, reps=5)
}
