% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_average_scores}
\alias{compute_average_scores}
\title{Get average scores}
\usage{
compute_average_scores(land1, land2, score)
}
\arguments{
\item{land1}{(numeric) vector with land ids of bullet 1}

\item{land2}{(numeric) vector with land ids of bullet 2}

\item{score}{numeric vector of scores to be summarized}
}
\value{
numeric vector of average scores. Length is the same as the number of land engraved areas on the bullets.
}
\description{
Get average scores
}
