% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{do_align}
\alias{do_align}
\title{Align two surface cross cuts using cross correlation}
\usage{
do_align(y1, y2)
}
\arguments{
\item{y1}{vector of striation marks (assuming equidistance between values)}

\item{y2}{vector of striation marks}
}
\value{
list consisting of a) the maximal cross correlation, b) the lag resulting in the highest cross correlation, 
and c) a vector of length y1 with aligned values of y2.
}
\description{
The first vector serves as a reference, the second vector is shifted, such that it aligns best with the first and has the same length as the first vector.
}

