% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p.R
\name{fortify}
\alias{fortify}
\title{Convert an x3p file into a data frame}
\usage{
fortify(x3p)
}
\arguments{
\item{x3p}{a file in x3p format as returned by function read_x3p}
}
\value{
data frame with variables x, y, and value
}
\description{
An x3p file consists of a list with header info and a 2d matrix with scan depths. 
fortify turns the matrix into a data frame, using the parameters of the header as necessary.
}
\examples{
data(br411)
br411_fort <- fortify(br411)
head(br411_fort)
}
