% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosssection.R
\name{get_bullet}
\alias{get_bullet}
\title{Deprecated function use get_crosscut}
\usage{
get_bullet(path, x = 243.75)
}
\arguments{
\item{path}{The path to the x3p file}

\item{x}{The crosscut value}
}
\description{
Deprecated function use get_crosscut
}
