% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley.R
\name{get_cor}
\alias{get_cor}
\title{Get correlation between two signatures}
\usage{
get_cor(b1, b2, window, b1.left, lag)
}
\arguments{
\item{b1}{dataframe}

\item{b2}{dataframe}

\item{window}{width of the window (in indices) to consider for matching}

\item{b1.left}{left index location of the matching window}

\item{lag}{integer lag for the second window}
}
\description{
Get correlation between two signatures
}
