% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosssection.R
\name{get_crosscut}
\alias{get_crosscut}
\title{Read a crosscut from a 3d surface file}
\usage{
get_crosscut(path = NULL, x = 243.75, bullet = NULL)
}
\arguments{
\item{path}{path to an x3p file. The path will only be considered, if bullet is not specified.}

\item{x}{level of the crosscut to be taken. If this level does not exist, the crosscut with the closest level is returned.}

\item{bullet}{alternative access to the surface measurements.}
}
\value{
data frame
}
\description{
Read a crosscut from a 3d surface file
}
