% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_D}
\alias{get_D}
\title{Compute the Euclidean distance between toolmarks}
\usage{
get_D(y1, y2, normalize = TRUE, resolution = 0.645)
}
\arguments{
\item{y1}{vector of equi-distant toolmark values}

\item{y2}{vector of equi-distant toolmark values}

\item{normalize}{should the result be normalized to 1000 microns (1 millimeter)? Defaults to TRUE.}

\item{resolution}{microns per pixel. Only used for normalization.}
}
\description{
Compute the Euclidean distance between two toolmark patterns. The striation patterns are not aligned before the distance is calculated.
}
