% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_features}
\alias{get_features}
\title{Get a feature vector for a pair of lands}
\usage{
get_features(res)
}
\arguments{
\item{res}{list of two aligned lands resulting from bulletGetMaxCMS}
}
\description{
Get a feature vector for a pair of lands
}
