% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves}
\alias{get_grooves}
\title{Find the grooves of a bullet land}
\usage{
get_grooves(bullet, method = "rollapply", smoothfactor = 15, adjust = 10,
  groove_cutoff = 400, mean_left = NULL, mean_right = NULL,
  mean_window = 100)
}
\arguments{
\item{bullet}{data frame with topological data in x-y-z format}

\item{method}{method to use for identifying grooves. Defaults to "rollapply"}

\item{smoothfactor}{The smoothing window to use - XXX the smoothing window seems to depend on the resolution at which the data has been collected.}

\item{adjust}{positive number to adjust the grooves - XXX should be expressed in microns rather than an index}

\item{groove_cutoff}{The index at which a groove cannot exist past - XXX this parameter should be expressed in microns rather than as an index to be able to properly deal with different resolutions}

\item{mean_left}{If provided, the location of the average left groove}

\item{mean_right}{If provided, the location of the average right groove}

\item{mean_window}{The window around the means to use}
}
\description{
Find the grooves of a bullet land
}
