% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_middle}
\alias{get_grooves_middle}
\title{Use the center of a crosscut}
\usage{
get_grooves_middle(bullet, middle = 75)
}
\arguments{
\item{bullet}{data frame with topological data in x-y-z format}

\item{middle}{middle percent to use for the identification}
}
\description{
Use the center of a crosscut
}
