% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_quadratic}
\alias{get_grooves_quadratic}
\title{Quadratic fit to find groove locations}
\usage{
get_grooves_quadratic(bullet, adjust)
}
\arguments{
\item{bullet}{data frame with topological data in x-y-z format}

\item{adjust}{positive number to adjust the grooves}
}
\description{
Use a robust fit of a quadratic curve to find groove locations
}
