% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooves.R
\name{get_grooves_rollapply}
\alias{get_grooves_rollapply}
\title{Using rollapply to find grooves in a crosscut}
\usage{
get_grooves_rollapply(bullet, smoothfactor = 15, adjust = 10,
  groove_cutoff = 400, mean_left = NULL, mean_right = NULL,
  mean_window = 100)
}
\arguments{
\item{bullet}{data frame with topological data in x-y-z format}

\item{smoothfactor}{The smoothing window to use}

\item{adjust}{positive number to adjust the grooves}

\item{groove_cutoff}{The index at which a groove cannot exist past}

\item{mean_left}{If provided, the location of the average left groove}

\item{mean_right}{If provided, the location of the average right groove}

\item{mean_window}{The window around the means to use}
}
\description{
Using rollapply to find grooves in a crosscut
}
