% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley.R
\name{get_lag}
\alias{get_lag}
\title{Get best lag for two vectors based on cross-correlation}
\usage{
get_lag(b1, b2, negperc = 10)
}
\arguments{
\item{b1}{vector of striation marks (assuming equidistance between values)}

\item{b2}{(smaller) vector of striation marks}

\item{negperc}{amount of lead that b2 can have compared to b1}
}
\value{
list of lag and correlation achieved. The plot shows b2 on b1
}
\description{
A small piece (b2) is matched to a much larger piece (b1). The lag gives the index location of the best match of b2 in b1.
This function is essentially just a wrapper for \code{my_ccf}, but adds a plot of the result for convenience.
}
