% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley.R
\name{get_lag_max_R}
\alias{get_lag_max_R}
\title{Get R Statistic for Chumbley matching}
\usage{
get_lag_max_R(b1, b2, window, b1.left)
}
\arguments{
\item{b1}{dataframe}

\item{b2}{dataframe}

\item{window}{width of the window (in indices) to consider for matching}

\item{b1.left}{left location of the matching window}
}
\description{
See Chumbley et al (2010)
}

