% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striations.R
\name{get_peaks}
\alias{get_peaks}
\title{Identify the location and the depth of peaks and heights at a crosscut}
\usage{
get_peaks(loessdata, column = "resid", smoothfactor = 35, striae = TRUE,
  window = TRUE)
}
\arguments{
\item{loessdata}{export from rollapply}

\item{column}{The column which should be smoothed}

\item{smoothfactor}{set to default of 35. Smaller values will pick up on smaller changes in the crosscut.}

\item{striae}{If TRUE, show the detected striae on the plot}

\item{window}{If TRUE, show the window of the striae on the plot}
}
\value{
list of several objects:
}
\description{
Identify the location and the depth of peaks and heights at a crosscut
}
