% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striations.R
\name{get_peaks_nist}
\alias{get_peaks_nist}
\title{Identify the location and the depth of peaks and heights at a crosscut}
\usage{
get_peaks_nist(loessdata, column = "resid", smoothfactor = 35)
}
\arguments{
\item{loessdata}{export from rollapply}

\item{column}{The column which should be smoothed}

\item{smoothfactor}{set to default of 35. Smaller values will pick up on smaller changes in the crosscut.}
}
\value{
list of several objects:
}
\description{
Identify the location and the depth of peaks and heights at a crosscut
}

