% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{getCircle}
\alias{getCircle}
\title{Estimate center and radius}
\usage{
getCircle(x, y)
}
\arguments{
\item{x}{numeric vector of values}

\item{y}{numeric vector of values}
}
\value{
three dimensional vector of the circle center (x0, y0) and the radius
}
\description{
Assuming the variables x and y are describing points located on a circle, the function uses a likelihood approach to estimate center and radius of the circle.
}
