% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image.x3p}
\alias{image.x3p}
\title{Plot x3p object as an image}
\usage{
\method{image}{x3p}(x, file = NULL, col = "#cd7f32", ...)
}
\arguments{
\item{x}{x3p object}

\item{file}{file name for saving, if file is NULL the opengl device stays open}

\item{col}{color specification}

\item{...}{not used}
}
\description{
Plot x3p object as an image
}
\examples{
data(br411)
image(br411, file = "br411.png")
}
