% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{maxCMS}
\alias{maxCMS}
\title{Number of maximum consecutively matching striae}
\usage{
maxCMS(match)
}
\arguments{
\item{match}{is a Boolean vector of matches/non-matches}
}
\value{
an integer value of the maximum number of consecutive matches
}
\description{
Number of maximum consecutively matching striae
}
\examples{
x <- rbinom(100, size = 1, prob = 1/3) 
CMS(x == 1) # expected value for longest match is 3
maxCMS(x==1)
}
