% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{predCircle}
\alias{predCircle}
\title{Estimate predictions and residuals for a circle fit of x and y}
\usage{
predCircle(x, y, resid.method = "response")
}
\arguments{
\item{x}{vector of numeric values}

\item{y}{vector of numeric values}

\item{resid.method}{character, one of "response" or "ortho"(gonal)}
}
\value{
data frame with predictions and residuals
}
\description{
estimate a circle, find predictive values and resiudals. depending on specification, vertical (regular) residuals or orthogonal residuals are computed.
}
