% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{predSmooth}
\alias{predSmooth}
\title{Estimate predictions and residuals for a smooth of x and y}
\usage{
predSmooth(x, y)
}
\arguments{
\item{x}{vector of numeric values}

\item{y}{vector of numeric values}
}
\value{
data frame with predictions and residuals
}
\description{
Fit a smooth line throught x and y, find predictive values and resiudals.
}
