% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\name{read_dat}
\alias{read_dat}
\title{Read a dat file and create an x3p file}
\format{list with header information and surface matrix}
\usage{
read_dat(path)
}
\arguments{
\item{path}{The file path to the dat file}
}
\description{
Read a dat file and create an x3p file
}
\examples{
\dontrun{
d1 <- read_dat("Br4 Bullet 4-1.dat")
}

}

