% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyapp.R
\name{runShinyapp}
\alias{runShinyapp}
\title{Shiny Apps: Find and launch apps bundled with bulletr}
\usage{
runShinyapp(name)
}
\arguments{
\item{name}{Name of the shiny app to be launched}
}
\description{
Shiny Apps: Find and launch apps bundled with bulletr
}
\examples{
\dontrun{
bulletr::runShinyapp('wrApp')
}

}
