% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossSection.R
\name{single_cc}
\alias{single_cc}
\title{Get crosscut at a specific level}
\usage{
single_cc(x, obj, obj_name = "")
}
\arguments{
\item{x}{level to use}

\item{obj}{x3p object (bullet)}

\item{obj_name}{name of x3p object (from file). Defaults to "".}
}
\description{
Get crosscut at a specific level
}
