% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/striations.R
\name{striation_identify}
\alias{striation_identify}
\title{Match striation marks across two cross sections based on previously identified peaks and valleys}
\usage{
striation_identify(lines1, lines2)
}
\arguments{
\item{lines1}{data frame as returned from get_peaks function. data frames are expected to have 
the following variables: xmin, xmax, group, type, bullet, heights}

\item{lines2}{data frame as returned from get_peaks function. data frames are expected to have 
the following variables: xmin, xmax, group, type, bullet, heights}
}
\value{
data frame of the same form as lines1 and lines2, but consisting of an additional variable of whether the striation marks are matches
}
\description{
Match striation marks across two cross sections based on previously identified peaks and valleys
}
