% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p.R
\name{unfortify_x3p}
\alias{unfortify_x3p}
\title{Convert a data frame into an x3p file}
\usage{
unfortify_x3p(df)
}
\arguments{
\item{df}{A data frame produced by fortify_x3p}
}
\value{
An x3p object
}
\description{
Convert a data frame into an x3p file
}
\examples{
data(br411)
br411_fort <- fortify_x3p(br411)
br411_unfort <- unfortify_x3p(br411_fort)
identical(br411_unfort, br411)
}
