% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_x3p2.R
\name{write_x3p}
\alias{write_x3p}
\title{Write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix as inputs}
\usage{
write_x3p(surface.matrix, file, general.info = NULL, feature.info = NULL,
  matrix.info = NULL, profiley = TRUE)
}
\arguments{
\item{surface.matrix}{Surface Matrix with the x y z values to be written (variable type: matrix)}

\item{file}{where should the file be stored?}

\item{general.info}{Setting the Values for the XML to a list}

\item{feature.info}{Setting the Values for the XML to a list}

\item{matrix.info}{Setting the Values for the XML to a list}

\item{profiley}{If FALSE, reorient the matrix to ensure a profile is taken is consistent with surface.matrix (input variable). The default value of Profiley is TRUE}
}
\description{
Write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix as inputs
}
\examples{
\dontrun{
 data(example_input_writex3p, envir=environment())
 write_x3p(surface.matrix=surface.matrix, file="out.x3p", profiley = FALSE)
 
}
}

