% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_x3p2.R
\name{write_x3p.x3p}
\alias{write_x3p.x3p}
\title{Write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix as inputs}
\usage{
\method{write_x3p}{x3p}(x, file, profiley = TRUE)
}
\arguments{
\item{x}{x3pobject}

\item{file}{where should the file be stored?}

\item{profiley}{If FALSE, reorient the matrix to ensure a profile is taken is consistent with surface.matrix (input variable). The default value of Profiley is TRUE}
}
\description{
Write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix as inputs
}
\examples{
\dontrun{
 # use all defaults:
 write_x3p(surface.matrix=surface.matrix, file="out.x3p", profiley = FALSE)
 
}
}
