% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_x3p_alt.R
\name{write_x3p_alt.default}
\alias{write_x3p_alt.default}
\title{Alternative Write function to write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix bulet_info as inputs}
\usage{
\method{write_x3p_alt}{default}(x, file, header.info = x$header.info,
  general.info = x$general.info, feature.info = x$feature.info,
  matrix.info = x$matrix.info, bullet_info = x$bullet_info,
  profiley = TRUE, template = NULL)
}
\arguments{
\item{x}{Surface Matrix with the x y z values to be written (variable type: matrix)}

\item{file}{where should the file be stored?}

\item{header.info}{header info of x3p}

\item{general.info}{Setting the Values for the XML to a list}

\item{feature.info}{Setting the Values for the XML to a list}

\item{matrix.info}{Setting the Values for the XML to a list}

\item{bullet_info}{xml file from x3p file}

\item{profiley}{If FALSE, reorient the matrix to ensure a profile is taken is consistent with surface.matrix (input variable). The default value of Profiley is TRUE}

\item{template}{path to an xml file with meta information to be saved as part of the x3p file.}
}
\description{
Alternative Write function to write an x3p file taking Lists: general.info, feature.info, matrix.info and matrix: surface.matrix bulet_info as inputs
}
\examples{
\dontrun{
 # use all defaults:
 write_x3p_alt(surface.matrix=surface.matrix, file="out.x3p", profiley = FALSE)

Examples to test various methods of using a specified template and without it.
a<- read_x3p(file.choose())
b<- read_dat(file.choose())
specified_template<- "./inst/shiny-examples/wrApp/defaultTemplateXML.xml"
# Write DAT file without specified template
write_x3p_alt(b, "test.x3p", profiley = T)
check<- read_x3p("./test.x3p")
# Write DAT file with specified template
write_x3p_alt(b, "test.x3p", profiley = T, template = specified_template)
check<- read_x3p("./test.x3p")
# Write x3p file with specified template
write_x3p_alt(a, "test.x3p", profiley = T, template = specified_template)
check<- read_x3p("./test.x3p")
# Write x3p file without specified template
write_x3p_alt(a, "test.x3p", profiley = T)
check<- read_x3p("./test.x3p")
 
}
}
