% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosssection.R
\name{bulletCheckCrossCut}
\alias{bulletCheckCrossCut}
\title{Identifying a reliable cross section}
\usage{
bulletCheckCrossCut(path, bullet = NULL, distance = 25, xlimits = c(50,
  500), minccf = 0.9, span = 0.03)
}
\arguments{
\item{path}{path to an x3p file}

\item{bullet}{If passed in, the actual bullet already loaded}

\item{distance}{positive numeric value indicating the distance between cross sections to use for a comparison}

\item{xlimits}{vector of values between which to check for cross sections in a stable region}

\item{minccf}{minimal value of cross correlation to indicate a stable region}

\item{span}{The span for the loess smooth function}
}
\description{
Should be changed: x should just indicate lower and upper limit. That is cleaner and should speed things up as well.
}
