% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chumbley.R
\name{get_lag_max_R}
\alias{get_lag_max_R}
\title{Get R Statistic for Chumbley matching}
\usage{
get_lag_max_R(b1, b2, window, b1.left)
}
\arguments{
\item{b1}{dataframe}

\item{b2}{dataframe}

\item{window}{width of the window (in indices) to consider for matching}

\item{b1.left}{left index of the matching window}
}
\value{
list of lag and correlation achieved. The plot shows b2 on b1.
}
\description{
See Chumbley et al (2010).
A small piece (b2) is matched to a much larger piece (b1). The lag gives the index location of the best match of b2 in b1.
}
