% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{plot_3d_land}
\alias{plot_3d_land}
\title{Plot a bullet land using plotly}
\usage{
plot_3d_land(path, bullet = NULL)
}
\arguments{
\item{path}{The path to the x3p file}

\item{bullet}{If not null, use this pre-loaded bullet}
}
\description{
Plot a bullet land using plotly
}
