% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_dat}
\alias{read_dat}
\title{Read a dat file and create an x3p file}
\format{list with header information and surface matrix}
\usage{
read_dat(path, profiley = TRUE, sample = 1)
}
\arguments{
\item{path}{The file path to the dat file}

\item{profiley}{are profiles on y?}

\item{sample}{1 in sample lines will be taken}
}
\description{
Read a dat file and create an x3p file
}
\examples{
\dontrun{
d1 <- read_dat("Br4 Bullet 4-1.dat")
d2 <- read_dat("L1.dat", profiley = FALSE)
}

}
